Resolves conflicts from #101 (generators.py rewrite) and #104 (sparse
ROI masks), both merged into dev after this branch was created off the
pre-merge sparse-roi-masks. Took dev's versions of code this branch
only carried a stale pre-merge copy of; kept this branch's own new
additions (NeuropilExtension, Suite2pRois.get_stats(), the from-stat
sparse.GCXS test coverage) on top.

Verified: full suite gives the same 18 pre-existing failures (env/
spikeinterface-version drift) as on unmodified neuropil-halo-extension
before this merge; 299 passed here vs. 272 there.

+++ src/photon_mosaic/core/tests/test_roianalyzer.py
@@ -105,7 +105,7 @@ class TestCreateMemory:
         assert analyzer.get_total_samples() == 100
 
     def test_get_dtype(self, analyzer):
-        assert analyzer.get_dtype() == np.float64
+        assert analyzer.get_dtype() == np.float32
 
     def test_is_read_only(self, analyzer):
         assert not analyzer.is_read_only()
